+++ src/Service.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace WeasyPrint;
 
 use Illuminate\Config\Repository;
+use Illuminate\Container\Container;
 use Illuminate\Contracts\Support\Renderable;
 use Rockett\Pipeline\Contracts\PipelineContract;
 use Rockett\Pipeline\Pipeline;
@@ -23,14 +24,15 @@ class Service implements Factory
   private Config $config;
   private Source $source;
 
-  public function __construct(Repository $config)
-  {
+  public function __construct(
+    Repository $config,
+  ) {
     $this->config = new Config(...$config->get('weasyprint'));
   }
 
   public static function instance(): self
   {
-    return app(Factory::class);
+    return Container::getInstance()->make(Factory::class);
   }
 
   public function getWeasyPrintVersion(): string